Bug Fixes: Several small bug fixes
Window was undefined prior to first adapt.
Increased space in a few arrays
length should be int, not sure how it worked before

+++ src/wlcsim/initcond.f03
@@ -58,7 +58,7 @@ real(dp) center(3)
 real(dp) RloopVec(3)
 real(dp) perpVec(3)
 real(dp) trash(3)
-real(dp) length
+integer length
 integer otherEnd
 real(dp) nloops
